fix: separating handle error functions
