+++ src/v1/destinations/iterable/strategies/track-identify.ts
@@ -1,7 +1,10 @@
 import { BaseStrategy } from './base';
-import { IterableBulkProxyInput } from '../types';
+import { GenericProxyHandlerInput, IterableBulkProxyInput } from '../types';
 import { checkIfEventIsAbortableAndExtractErrorMessage } from '../utils';
 import { DeliveryJobState, DeliveryV1Response } from '../../../../types';
+import { TransformerProxyError } from '../../../../v0/util/errorTypes';
+import { getDynamicErrorType } from '../../../../adapters/utils/networkUtils';
+import { TAG_NAMES } from '../../../../v0/util/tags';
 
 class TrackIdentifyStrategy extends BaseStrategy {
   handleSuccess(responseParams: IterableBulkProxyInput): DeliveryV1Response {
@@ -38,6 +41,28 @@ class TrackIdentifyStrategy extends BaseStrategy {
       destinationResponse,
       response: responseWithIndividualEvents,
     };
+  }
+
+  handleError(responseParams: GenericProxyHandlerInput): void {
+    const { destinationResponse, rudderJobMetadata } = responseParams;
+    const { response, status } = destinationResponse;
+    const responseMessage = response.params || response.msg || response.message;
+    const errorMessage = JSON.stringify(responseMessage) || 'unknown error format';
+
+    const responseWithIndividualEvents = rudderJobMetadata.map((metadata) => ({
+      statusCode: status,
+      metadata,
+      error: errorMessage,
+    }));
+
+    throw new TransformerProxyError(
+      `ITERABLE: Error transformer proxy during ITERABLE response transformation. ${errorMessage}`,
+      status,
+      { [TAG_NAMES.ERROR_TYPE]: getDynamicErrorType(status) },
+      destinationResponse,
+      '',
+      responseWithIndividualEvents,
+    );
   }
 }
